Properly setup building and packaging of the lib.
Splitted the lib and the executable.
Added some macrodefs to properly link the library.
Replaced int in the signature with a enum. Should have the same size and not break compatibility, but increase clarity. Also changed its values hardcoded into the code to the names in the enum.
Added support of CMake.
Added CPack packaging.
Added license as a separate file.
Added ReadMe.md with some text.